renamed annotation/tool_ids.ts to better fit its contents, removed unnecessary comments
