+++ src/layer/annotation/tool_state.ts
