+++ src/ui/annotations.ts
@@ -56,7 +56,7 @@ import type { CoordinateSpace } from "#src/coordinate_transform.js";
 import {
   SELECT_NEXT_ANNOTATION_TOOL_ID,
   SELECT_PREVIOUS_ANNOTATION_TOOL_ID,
-} from "#src/layer/annotation/tool_ids.js";
+} from "#src/layer/annotation/tool_state.js";
 import type { MouseSelectionState, UserLayer } from "#src/layer/index.js";
 import type { LoadedDataSubsource } from "#src/layer/layer_data_source.js";
 import type { ChunkTransformParameters } from "#src/render_coordinate_transform.js";
